requirements.txt
discord.py>=2.5.2
python-dotenv>=1.1.0
requests>=2.32.3
beautifulsoup4>=4.13.4
selenium>=4.31.0
webdriver-manager>=4.0.2
aiohttp>=3.9.3
